Add a namedtuple RefreshTokenInfo to be returned by TokenStore.lookup to allow it to be stubbed without depending on ndb.Model. In Cloud NDB, ndb.Model creation requires a ndb client context regardless of if  or  on it are used.
