+++ refresh_token.py
@@ -1,11 +1,3 @@
+from collections import namedtuple
 from google.appengine.ext import ndb
 
-
-class RefreshToken(ndb.Model):
-    token = ndb.StringProperty()
-    issued_at = ndb.DateTimeProperty()
-    username = ndb.StringProperty()
-
-    @classmethod
-    def kind_name(cls):
-        return cls.__name__
