+++ tests/datastore_emulator/token_store_test.py
@@ -1,7 +1,7 @@
 import unittest
 
 from google.appengine.ext import ndb
-from token_store import TokenStore
+from token_store import TokenStore, RefreshToken
 from datetime import datetime
 
 provider_name = "test"
@@ -30,8 +30,7 @@ class TokenStoreTestCase(unittest.TestCase):
     def test_save(self):
         token_store = TokenStore()
         self.assertIsNone(self.key.get())
-        result_key = token_store.save(self.user_id, self.token_str, self.issued_at, self.username, provider_name)
-        self.assertEqual(result_key, self.key)
+        token_store.save(self.user_id, self.token_str, self.issued_at, self.username, provider_name)
         saved_token = self.key.get()
         self.assertIsNotNone(saved_token)
         self.assertEqual(self.token_str, saved_token.token)
@@ -45,3 +44,18 @@ class TokenStoreTestCase(unittest.TestCase):
         self.assertEqual(self.token_str, persisted_token.token)
         self.assertEqual(self.issued_at, persisted_token.issued_at)
         self.assertEqual(self.username, persisted_token.username)
+
+class RefreshTokenTestCase(unittest.TestCase):
+
+    def test_kind_name(self):
+        self.assertEqual("RefreshToken", RefreshToken.kind_name())
+
+    def test_properties(self):
+        user_id = "123456"
+        token_str = "foobarbaz"
+        issued_at = datetime.now()
+        username = "bob@email-address.gov"
+        refresh_token = RefreshToken(id=user_id, token=token_str, issued_at=issued_at, username=username)
+        self.assertEqual(user_id, refresh_token.key.id())
+        self.assertEqual(token_str, refresh_token.token)
+        self.assertEqual(issued_at, refresh_token.issued_at)
